Deduplicate entity-events-v1 mixins
