+++ legacy-fabric-entity-events-v1/1.8.9/src/main/resources/legacy-fabric-entity-events-v1.mixins.json
@@ -4,9 +4,8 @@
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "EntityMixin",
-    "LivingEntityMixin",
-    "PlayerManagerMixin",
-    "ServerPlayerEntityMixin"
+    "versionned.PlayerManagerMixin",
+    "versionned.ServerPlayerEntityMixin"
   ],
   "injectors": {
     "defaultRequire": 1
