+++ legacy-fabric-entity-events-v1/1.9.4/src/main/java/net/legacyfabric/fabric/mixin/entity/event/versionned/PlayerManagerMixin.java
@@ -15,22 +15,16 @@
  * limitations under the License.
  */
 
-package net.legacyfabric.fabric.mixin.entity.event;
-
-import java.util.Iterator;
+package net.legacyfabric.fabric.mixin.entity.event.versionned;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.server.PlayerManager;
-import net.minecraft.server.world.ServerWorld;
 
-import net.legacyfabric.fabric.api.entity.event.v1.ServerEntityWorldChangeEvents;
 import net.legacyfabric.fabric.api.entity.event.v1.ServerPlayerEvents;
 
 @Mixin(PlayerManager.class)
@@ -38,14 +32,5 @@ abstract class PlayerManagerMixin {
 	@Inject(method = "respawnPlayer", at = @At("TAIL"))
 	private void afterRespawn(ServerPlayerEntity oldPlayer, int dimension, boolean alive, CallbackInfoReturnable<ServerPlayerEntity> cir) {
 		ServerPlayerEvents.AFTER_RESPAWN.invoker().afterRespawn(oldPlayer, cir.getReturnValue(), oldPlayer.server.getWorld(dimension), alive);
-	}
-
-	/**
-	 * This is called by both "moveToWorld" and "teleport".
-	 * So this is suitable to handle the after event from both call sites.
-	 */
-	@Inject(method = "teleportToDimension", at = @At("TAIL"), locals = LocalCapture.CAPTURE_FAILHARD)
-	private void afterWorldChanged(ServerPlayerEntity player, int dimension, CallbackInfo ci, int i, ServerWorld serverWorld, ServerWorld serverWorld2, Iterator iterator) {
-		ServerEntityWorldChangeEvents.AFTER_PLAYER_CHANGE_WORLD.invoker().afterChangeWorld(player, serverWorld, serverWorld2);
 	}
 }
