+++ legacy-fabric-entity-events-v1/common/src/main/java/net/legacyfabric/fabric/mixin/entity/event/LivingEntityMixin.java
