+++ legacy-fabric-entity-events-v1/common/src/main/java/net/legacyfabric/fabric/mixin/entity/event/PlayerManagerMixin.java
@@ -17,20 +17,29 @@
 
 package net.legacyfabric.fabric.mixin.entity.event;
 
+import java.util.Iterator;
+
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.damage.DamageSource;
+import net.minecraft.entity.player.ServerPlayerEntity;
+import net.minecraft.server.PlayerManager;
+import net.minecraft.server.world.ServerWorld;
 
-import net.legacyfabric.fabric.api.entity.event.v1.ServerEntityCombatEvents;
+import net.legacyfabric.fabric.api.entity.event.v1.ServerEntityWorldChangeEvents;
 
-@Mixin(LivingEntity.class)
-abstract class LivingEntityMixin {
-	@Inject(method = "onKilled", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/Entity;onKilledOther(Lnet/minecraft/entity/LivingEntity;)V", shift = At.Shift.AFTER))
-	public void onEntityKilledOther(DamageSource source, CallbackInfo ci) {
-		ServerEntityCombatEvents.AFTER_KILLED_OTHER_ENTITY.invoker().afterKilledOtherEntity(source.getAttacker(), (LivingEntity) (Object) this);
+@Mixin(PlayerManager.class)
+public class PlayerManagerMixin {
+
+	/**
+	 * This is called by both "moveToWorld" and "teleport".
+	 * So this is suitable to handle the after event from both call sites.
+	 */
+	@Inject(method = "teleportToDimension", at = @At("TAIL"), locals = LocalCapture.CAPTURE_FAILHARD)
+	private void afterWorldChanged(ServerPlayerEntity player, int dimension, CallbackInfo ci, int i, ServerWorld serverWorld, ServerWorld serverWorld2, Iterator iterator) {
+		ServerEntityWorldChangeEvents.AFTER_PLAYER_CHANGE_WORLD.invoker().afterChangeWorld(player, serverWorld, serverWorld2);
 	}
 }
