+++ legacy-fabric-entity-events-v1/common/src/main/resources/fabric.mod.json
@@ -19,7 +19,9 @@
   "depends": {
     "minecraft": "${minecraft_version}"
   },
-  "mixins": [],
+  "mixins": [
+    "legacy-fabric-entity-events-v1-common.mixins.json"
+  ],
   "custom": {
     "modmenu": {
       "badges": [ "library" ],
